# kaushaldrishti/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "KaushalDrishti API",
    "description": "AI-enabled Labour Market Intelligence System (LMIS) for MSDE. Provides district-level labour demand indices, supply forecasts, gap analysis, early warnings, and policy scenario simulation.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/webhooks": {
      "get": {
        "tags": [
          "webhooks"
        ],
        "summary": "List Webhooks",
        "description": "List all registered webhook listeners.",
        "operationId": "list_webhooks_api_v1_webhooks_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/WebhookResponse"
                  },
                  "type": "array",
                  "title": "Response List Webhooks Api V1 Webhooks Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "webhooks"
        ],
        "summary": "Subscribe Webhook",
        "description": "Register a webhook listener for early warning flag changes.",
        "operationId": "subscribe_webhook_api_v1_webhooks_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WebhookSubscribeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/webhooks/test": {
      "post": {
        "tags": [
          "webhooks"
        ],
        "summary": "Deliver Test Event",
        "description": "Simulate and deliver an alert state change event to test the webhook pipeline.",
        "operationId": "deliver_test_event_api_v1_webhooks_test_post",
        "parameters": [
          {
            "name": "webhook_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Specific webhook ID to test",
              "title": "Webhook Id"
            },
            "description": "Specific webhook ID to test"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookTestDeliveryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WebhookResponse": {
        "properties": {
          "webhook_id": {
            "type": "string",
            "title": "Webhook Id"
          },
          "target_url": {
            "type": "string",
            "title": "Target Url"
          },
          "events": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Events"
          },
          "state_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "State Code"
          },
          "district_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "District Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "webhook_id",
          "target_url",
          "events",
          "status",
          "created_at"
        ],
        "title": "WebhookResponse"
      },
      "WebhookSubscribeRequest": {
        "properties": {
          "target_url": {
            "type": "string",
            "title": "Target Url"
          },
          "events": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Events",
            "default": [
              "alert.created",
              "alert.escalated",
              "alert.de_escalated"
            ]
          },
          "secret_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Secret Token"
          },
          "state_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "State Code"
          },
          "district_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "District Id"
          }
        },
        "type": "object",
        "required": [
          "target_url"
        ],
        "title": "WebhookSubscribeRequest"
      },
      "WebhookTestDeliveryResponse": {
        "properties": {
          "webhook_id": {
            "type": "string",
            "title": "Webhook Id"
          },
          "delivered": {
            "type": "boolean",
            "title": "Delivered"
          },
          "event_type": {
            "type": "string",
            "title": "Event Type"
          },
          "payload": {
            "type": "object",
            "title": "Payload"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "webhook_id",
          "delivered",
          "event_type",
          "payload",
          "message"
        ],
        "title": "WebhookTestDeliveryResponse"
      }
    }
  }
}